Add failing test case.
It does not compile without changes.

The idea behind the test to using different domains and different
schedulers.

These schedulers are waited in a transfer_when_all function. With this
sender the final result should be transfered to the given domain.

Right now it doesn't compiles because the senders has no common domain.