Make transfer_when_all domain independent.
The idea is that transfer_when_all actually doesn't necessary to expect
senders from the same domain, because it's gonna be anyways transfered.

So, instead of letting unspecified the domain where the when_all runs
the transfer_when_all as a last step transfers all the senders to the
final_domain and then it will execute the when_all.

This fixes the compile issue and makes easier the work with different
domains.

It has a side effect while one of the test needed to be changed. The
when_all in this situation will be executed on the domain what was given
as a parameter.